requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
pandas
black
numpy
jupyter
sklearn
matplotlib
seaborn
flake8
xgboost
catboost
tpot
python-dotenv>=0.5.1
jupyter
